chore: split AgentBuilderView into mulitple components
